+++ internal/utils/pagination.go
@@ -1,11 +1,11 @@
 package utils
 
-func GetPagination(limit, page int32) (newLimit, offset int) {
+func GetPagination(limit, page int32) (newLimit, offset int64) {
 	if page > 0 {
-		offset = int(page - 1)
+		offset = int64(page - 1)
 	}
 	if limit > 0 {
-		newLimit = int(limit)
+		newLimit = int64(limit)
 	}
 	offset = offset * newLimit
 	return
